refactor: apply Sp/Spp naming convention to species components and actions
Single-species components get Sp prefix, multi-species get Spp.
Removes ambiguity since "species" is identical singular/plural.
